// src/Lineage_old/types.generated.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Long: number;
};
// ...
export type BrowsableEntity = {
  /** The browse paths corresponding to an entity. If no Browse Paths have been generated before, this will be null. */
  browsePaths?: Maybe<Array<BrowsePath>>;
};
// ...
export type BrowsePath = {
  __typename?: 'BrowsePath';
  /** The components of the browse path */
  path: Array<Scalars['String']>;
};

export type BrowsePathEntry = {
  __typename?: 'BrowsePathEntry';
  /**
   * An optional entity associated with this browse entry. This will usually be a container entity.
   * If this entity is not populated, the name must be used.
   */
  entity?: Maybe<Entity>;
  /** The path name of a group of browse results */
  name: Scalars['String'];
};

/** A hierarchical entity path V2 */
export type BrowsePathV2 = {
  __typename?: 'BrowsePathV2';
  /** The components of the browse path */
  path: Array<BrowsePathEntry>;
};
// ...
export type Chart = BrowsableEntity & Entity & EntityWithRelationships & {
  __typename?: 'Chart';
  /**
   * Experimental API.
   * For fetching extra entities that do not have custom UI code yet
   */
  aspects?: Maybe<Array<RawAspect>>;
  /** The browse path V2 corresponding to an entity. If no Browse Paths V2 have been generated before, this will be null. */
  browsePathV2?: Maybe<BrowsePathV2>;
  /** The browse paths corresponding to the chart. If no Browse Paths have been generated before, this will be null. */
  browsePaths?: Maybe<Array<BrowsePath>>;
  /** An id unique within the charting tool */
  chartId: Scalars['String'];
  /** The parent container in which the entity resides */
  container?: Maybe<Container>;
  /** The specific instance of the data platform that this entity belongs to */
  dataPlatformInstance?: Maybe<DataPlatformInstance>;
  /** The deprecation status of the chart */
  deprecation?: Maybe<Deprecation>;
  /** The Domain associated with the Chart */
  domain?: Maybe<DomainAssociation>;
  /**
   * Deprecated, use editableProperties field instead
   * Additional read write information about the Chart
   * @deprecated Field no longer supported
   */
  editableInfo?: Maybe<ChartEditableProperties>;
  /** Additional read write properties about the Chart */
  editableProperties?: Maybe<ChartEditableProperties>;
  /** Embed information about the Chart */
  embed?: Maybe<Embed>;
  /** Whether or not this entity exists on DataHub */
  exists?: Maybe<Scalars['Boolean']>;
  /** The forms associated with the Dataset */
  forms?: Maybe<Forms>;
  /**
   * Deprecated, use tags instead
   * The structured tags associated with the chart
   * @deprecated Field no longer supported
   */
  globalTags?: Maybe<GlobalTags>;
  /** The structured glossary terms associated with the dashboard */
  glossaryTerms?: Maybe<GlossaryTerms>;
  /** Experimental! The resolved health statuses of the asset */
  health?: Maybe<Array<Health>>;
  /** Incidents associated with the Chart */
  incidents?: Maybe<EntityIncidentsResult>;
  /**
   * Deprecated, use properties field instead
   * Additional read only information about the chart
   * @deprecated Field no longer supported
   */
  info?: Maybe<ChartInfo>;
  /** Input fields to power the chart */
  inputFields?: Maybe<InputFields>;
  /** References to internal resources related to the dashboard */
  institutionalMemory?: Maybe<InstitutionalMemory>;
  /** The timestamp for the last time this entity was ingested */
  lastIngested?: Maybe<Scalars['Long']>;
  /** Edges extending from this entity grouped by direction in the lineage graph */
  lineage?: Maybe<EntityLineageResult>;
  /** Ownership metadata of the chart */
  ownership?: Maybe<Ownership>;
  /** Recursively get the lineage of containers for this entity */
  parentContainers?: Maybe<ParentContainersResult>;
  /** Standardized platform urn where the chart is defined */
  platform: DataPlatform;
  /** Privileges given to a user relevant to this entity */
  privileges?: Maybe<EntityPrivileges>;
  /** Additional read only properties about the Chart */
  properties?: Maybe<ChartProperties>;
  /** Info about the query which is used to render the chart */
  query?: Maybe<ChartQuery>;
  /** Granular API for querying edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /**
   * Not yet implemented.
   *
   * Experimental - Summary operational & usage statistics about a Chart
   */
  statsSummary?: Maybe<ChartStatsSummary>;
  /** Status metadata of the chart */
  status?: Maybe<Status>;
  /** Structured properties about this asset */
  structuredProperties?: Maybe<StructuredProperties>;
  /** Sub Types that this entity implements */
  subTypes?: Maybe<SubTypes>;
  /** The tags associated with the chart */
  tags?: Maybe<GlobalTags>;
  /**
   * The chart tool name
   * Note that this field will soon be deprecated in favor a unified notion of Data Platform
   */
  tool: Scalars['String'];
  /** A standard Entity Type */
  type: EntityType;
  /** The primary key of the Chart */
  urn: Scalars['String'];
};
// ...
export type ChartEditableProperties = {
  __typename?: 'ChartEditableProperties';
  /** Description of the Chart */
  description?: Maybe<Scalars['String']>;
};
// ...
export type DataPlatformInstance = Entity & {
  __typename?: 'DataPlatformInstance';
  /** The deprecation status of the data platform instance */
  deprecation?: Maybe<Deprecation>;
  /** The platform instance id */
  instanceId: Scalars['String'];
  /** References to internal resources related to the data platform instance */
  institutionalMemory?: Maybe<InstitutionalMemory>;
  /** Ownership metadata of the data platform instance */
  ownership?: Maybe<Ownership>;
  /** Name of the data platform */
  platform: DataPlatform;
  /** Additional read only properties associated with a data platform instance */
  properties?: Maybe<DataPlatformInstanceProperties>;
  /** Edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** Status metadata of the container */
  status?: Maybe<Status>;
  /** Tags used for searching the data platform instance */
  tags?: Maybe<GlobalTags>;
  /** A standard Entity Type */
  type: EntityType;
  /** Urn of the data platform */
  urn: Scalars['String'];
};
// ...
export type Deprecation = {
  __typename?: 'Deprecation';
  /** The user who will be credited for modifying this deprecation content */
  actor?: Maybe<Scalars['String']>;
  /** The hydrated user who will be credited for modifying this deprecation content */
  actorEntity?: Maybe<Entity>;
  /** The time user plan to decommission this entity */
  decommissionTime?: Maybe<Scalars['Long']>;
  /** Whether the entity has been deprecated by owner */
  deprecated: Scalars['Boolean'];
  /** Additional information about the entity deprecation plan */
  note?: Maybe<Scalars['String']>;
  /** The optional replacement entity */
  replacement?: Maybe<Entity>;
};
// ...
export type DomainAssociation = {
  __typename?: 'DomainAssociation';
  /** Reference back to the tagged urn for tracking purposes e.g. when sibling nodes are merged together */
  associatedUrn: Scalars['String'];
  /** The domain related to the assocaited urn */
  domain: Domain;
};
// ...
export type Embed = {
  __typename?: 'Embed';
  /** A URL which can be rendered inside of an iframe. */
  renderUrl?: Maybe<Scalars['String']>;
};

/** A top level Metadata Entity */
export type Entity = {
  /** List of relationships between the source Entity and some destination entities with a given types */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** A standard Entity Type */
  type: EntityType;
  /** A primary key of the Metadata Entity */
  urn: Scalars['String'];
};
// ...
export enum EntityType {
  /** A DataHub Access Token */
  AccessToken = 'ACCESS_TOKEN',
  /** A DataHub Assertion */
  Assertion = 'ASSERTION',
  /** A Business Attribute */
  BusinessAttribute = 'BUSINESS_ATTRIBUTE',
  /** The Chart Entity */
  Chart = 'CHART',
  /** A container of Metadata Entities */
  Container = 'CONTAINER',
  /** The CorpGroup Entity */
  CorpGroup = 'CORP_GROUP',
  /** The CorpUser Entity */
  CorpUser = 'CORP_USER',
  /** A Custom Ownership Type */
  CustomOwnershipType = 'CUSTOM_OWNERSHIP_TYPE',
  /** The Dashboard Entity */
  Dashboard = 'DASHBOARD',
  /** A connection to an external source. */
  DatahubConnection = 'DATAHUB_CONNECTION',
  /** A DataHub Policy */
  DatahubPolicy = 'DATAHUB_POLICY',
  /** A DataHub Role */
  DatahubRole = 'DATAHUB_ROLE',
  /** A DataHub View */
  DatahubView = 'DATAHUB_VIEW',
  /** The Dataset Entity */
  Dataset = 'DATASET',
  /** A data contract */
  DataContract = 'DATA_CONTRACT',
  /** The Data Flow (or Data Pipeline) Entity, */
  DataFlow = 'DATA_FLOW',
  /** The Data Job (or Data Task) Entity */
  DataJob = 'DATA_JOB',
  /** The DataPlatform Entity */
  DataPlatform = 'DATA_PLATFORM',
  /** Data Platform Instance Entity */
  DataPlatformInstance = 'DATA_PLATFORM_INSTANCE',
  /** An instance of an individual run of a data job or data flow */
  DataProcessInstance = 'DATA_PROCESS_INSTANCE',
  /** A Data Product */
  DataProduct = 'DATA_PRODUCT',
  /**
   * "
   * A data type registered to DataHub
   */
  DataType = 'DATA_TYPE',
  /** A Domain containing Metadata Entities */
  Domain = 'DOMAIN',
  /**
   * "
   * A type of entity registered to DataHub
   */
  EntityType = 'ENTITY_TYPE',
  /** The ERModelRelationship Entity */
  ErModelRelationship = 'ER_MODEL_RELATIONSHIP',
  /** A DataHub ExecutionRequest */
  ExecutionRequest = 'EXECUTION_REQUEST',
  /**
   * "
   * A form entity on entities
   */
  Form = 'FORM',
  /** The Glossary Node Entity */
  GlossaryNode = 'GLOSSARY_NODE',
  /** The Glossary Term Entity */
  GlossaryTerm = 'GLOSSARY_TERM',
  /** A DataHub incident - SaaS only */
  Incident = 'INCIDENT',
  /** A DataHub Managed Ingestion Source */
  IngestionSource = 'INGESTION_SOURCE',
  /** The ML Feature Entity */
  Mlfeature = 'MLFEATURE',
  /** ML Feature Table Entity */
  MlfeatureTable = 'MLFEATURE_TABLE',
  /** The ML Model Entity */
  Mlmodel = 'MLMODEL',
  /** The MLModelGroup Entity */
  MlmodelGroup = 'MLMODEL_GROUP',
  /** The ML Primary Key Entity */
  MlprimaryKey = 'MLPRIMARY_KEY',
  /** The Notebook Entity */
  Notebook = 'NOTEBOOK',
  /** Another entity type - refer to a provided entity type urn. */
  Other = 'OTHER',
  /** A DataHub Post */
  Post = 'POST',
  /** A dataset query */
  Query = 'QUERY',
  /**
   * "
   * A type of entity that is restricted to the user
   */
  Restricted = 'RESTRICTED',
  /**
   * "
   * A Role from an organisation
   */
  Role = 'ROLE',
  /** A Schema Field */
  SchemaField = 'SCHEMA_FIELD',
  /**
   * "
   * An structured property on entities
   */
  StructuredProperty = 'STRUCTURED_PROPERTY',
  /** The Tag Entity */
  Tag = 'TAG',
  /** A DataHub Test */
  Test = 'TEST',
  /** A set of versioned entities, representing a single source / logical entity over time */
  VersionSet = 'VERSION_SET'
}
// ...
export type EntityWithRelationships = {
  /** Edges extending from this entity grouped by direction in the lineage graph */
  lineage?: Maybe<EntityLineageResult>;
  /** Granular API for querying edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** A standard Entity Type */
  type: EntityType;
  /** A primary key associated with the Metadata Entity */
  urn: Scalars['String'];
};
// ...
export type GlobalTags = {
  __typename?: 'GlobalTags';
  /** The set of tags attached to the Metadata Entity */
  tags?: Maybe<Array<TagAssociation>>;
};
// ...
export type RawAspect = {
  __typename?: 'RawAspect';
  /** The name of the aspect */
  aspectName: Scalars['String'];
  /** JSON string containing the aspect's payload */
  payload?: Maybe<Scalars['String']>;
  /** Details for the frontend on how the raw aspect should be rendered */
  renderSpec?: Maybe<AspectRenderSpec>;
};
// ...
export type SearchResult = {
  __typename?: 'SearchResult';
  /** The resolved DataHub Metadata Entity matching the search query */
  entity: Entity;
  /** Additional properties about the search result. Used for rendering in the UI */
  extraProperties?: Maybe<Array<ExtraProperty>>;
  /** Insights about why the search result was matched */
  insights?: Maybe<Array<SearchInsight>>;
  /** Matched field hint */
  matchedFields: Array<MatchedField>;
};